Add files via upload
Business Context:

In today's digital era, customer reviews are pivotal for e-commerce platforms and online service providers, serving as a barometer for customer satisfaction, highlighting areas for improvement, and guiding business decisions. The intricate process of Aspect-based Sentiment Analysis is instrumental in dissecting these reviews, offering a granular understanding of customer sentiments towards various facets of a product or service. This nuanced approach enables businesses to pinpoint specific elements—be it the screen, keyboard, or customer service of a laptop—that may need enhancement.

Moreover, the advent of Large Language Models (LLMs) has revolutionized the way businesses approach the summarization of customer reviews and case briefs. By leveraging LLMs, companies can swiftly and accurately distill the essence of customer feedback, streamlining the process of understanding overall sentiments and facilitating the refinement of product offerings. This swift synthesis not only conserves time and resources but also plays a critical role in enhancing customer satisfaction, driving sales, and boosting revenue.

Additionally, this project will delve into the fine-tuning of models to elevate the quality of summaries provided by LLMs. This involves training models on domain-specific data to enhance their ability to generate more precise and relevant summaries, thus improving the utility of these summaries in real-world applications.

Embarking on this final project on Customer Review Aspect-based Classification & Summarization, coupled with an emphasis on model fine-tuning for enhanced summarization, equips you with invaluable skills applicable to real-world business contexts. Through hands-on experience with code and implementation specifics, you'll gain the proficiency to construct such solutions using open-source LLMs. This experience will not only serve as a compelling Proof-of-Concept but also pave the way for the deployment and productionization of these cutting-edge solutions in business environments.
Project Objective

Develop a Generative AI application using a Large Language Model to automate the Aspect-based Classification and Summarization of Customer Reviews received by a business. The application will aim to predict review sentiment, aspect-based sentiment and generate summaries of these customer reviews, and will evaluate the performance of the LLM at these tasks through various evaluation schemes.